requirements.txt
openai==1.61.1
Pillow==11.1.0
python-dotenv==1.0.1
Requests==2.32.3
streamlit==1.31.0

